requirements.txt
asgiref==3.11.1
colorama==0.4.6
dj-database-url==3.1.2
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
iniconfig==2.3.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
